+++ crates/biome_js_analyze/tests/specs/nursery/noUselessLoneBlockStatements/invalid.cjs
@@ -1,0 +1,17 @@
+{}
+
+{
+ aLabel: {
+ }
+}
+
+{
+ function foo() {}
+}
+
+{
+ // a comment
+ function foo() {}
+}
+
+