+++ crates/biome_js_analyze/tests/specs/nursery/noUselessLoneBlockStatements/invalidModule.js
@@ -1,0 +1,47 @@
+export {}
+
+{}
+
+if (foo) {
+ bar();
+ {
+  baz();
+ }
+}
+
+if (foo) {
+ // this is bar
+ bar();
+ {
+  // this is baz
+  baz();
+ }
+}
+
+function bar() {
+ {
+  baz();
+ }
+}
+
+{
+ function foo() {}
+}
+
+{
+ // a comment
+ function foo() {}
+}
+
+{
+ aLabel: {
+ }
+}
+
+class C {
+ static {
+  {
+   foo();
+  }
+ }
+}