+++ crates/biome_js_analyze/tests/specs/nursery/useAriaActivedescendantWithTabindex/invalid.js
@@ -1,0 +1,1 @@
+<div aria-activedescendant={someID} />;