+++ packages/@biomejs/backend-jsonrpc/src/workspace.ts
@@ -718,6 +718,10 @@ export interface Nursery {
 	 */
 	noGlobalIsNan?: RuleConfiguration;
 	/**
+	 * Enforce that non-interactive ARIA roles are not assigned to interactive HTML elements.
+	 */
+	noInteractiveElementToNoninteractiveRole?: RuleConfiguration;
+	/**
 	 * Disallow new operators with global non-constructor functions.
 	 */
 	noInvalidNewBuiltin?: RuleConfiguration;
@@ -738,6 +742,10 @@ export interface Nursery {
 	 */
 	noUselessElse?: RuleConfiguration;
 	/**
+	 * Disallow unnecessary nested block statements.
+	 */
+	noUselessLoneBlockStatements?: RuleConfiguration;
+	/**
 	 * Disallow the use of void operators, which is not a familiar operator.
 	 */
 	noVoid?: RuleConfiguration;
@@ -745,6 +753,10 @@ export interface Nursery {
 	 * It enables the recommended rules for this group
 	 */
 	recommended?: boolean;
+	/**
+	 * Enforce that tabIndex is assigned to non-interactive HTML elements with aria-activedescendant.
+	 */
+	useAriaActivedescendantWithTabindex?: RuleConfiguration;
 	/**
 	 * Use arrow functions over function expressions.
 	 */
@@ -758,7 +770,7 @@ export interface Nursery {
 	 */
 	useCollapsedElseIf?: RuleConfiguration;
 	/**
-	 * Enforce all dependencies are correctly specified.
+	 * Enforce all dependencies are correctly specified in a React hook.
 	 */
 	useExhaustiveDependencies?: RuleConfiguration;
 	/**
@@ -1340,12 +1352,15 @@ export type Category =
 	| "lint/nursery/noFallthroughSwitchClause"
 	| "lint/nursery/noGlobalIsFinite"
 	| "lint/nursery/noGlobalIsNan"
+	| "lint/nursery/noInteractiveElementToNoninteractiveRole"
 	| "lint/nursery/noInvalidNewBuiltin"
 	| "lint/nursery/noMisleadingInstantiator"
 	| "lint/nursery/noMisrefactoredShorthandAssign"
 	| "lint/nursery/noUnusedImports"
 	| "lint/nursery/noUselessElse"
+	| "lint/nursery/noUselessLoneBlockStatements"
 	| "lint/nursery/noVoid"
+	| "lint/nursery/useAriaActivedescendantWithTabindex"
 	| "lint/nursery/useArrowFunction"
 	| "lint/nursery/useAsConstAssertion"
 	| "lint/nursery/useBiomeSuppressionComment"